feat: my asset actions refactor
