+++ src/components/index.ts
@@ -2,8 +2,6 @@ export { default as Header } from './header/Header'
 export { default as Footer } from './footer/index'
 export { default as MyTable } from './my-table'
 export { default as SearchInput } from './search-input/SearchInput'
-export { default as MarketNftListCard } from './nft-card/MarketNftListCard'
-export { default as MyAssetNftListCard } from './nft-card/MyAssetNftListCard'
 export { default as Select } from './custom-select/CustomSelect'
 export { default as AsyncSelectCollection } from './custom-select/AsyncSelectCollection'
 export { default as BaseRateTable } from './base-rate-table/BaseRateTable'
@@ -21,5 +19,7 @@ export { default as NotFound } from './not-found/NotFound'
 export { default as NftMedia } from './nft-media/NftMedia'
 export { default as ScrollNumber } from './scroll-number/ScrollNumber'
 export { default as H5SecondaryHeader } from './h5-secondary-header/H5SecondaryHeader'
+export { default as MortgagedTag } from './tag/MortgagedTag'
+export { default as ListingTag } from './tag/ListingTag'
 
 export { type ColumnProps } from './my-table/index'
