+++ src/pages/buy-nfts/components/MarketNftListCard.tsx
